requirements.txt
fastapi
pydantic
pydantic-settings
pyjwt[crypto]
uvicorn[standard]
websockets
# ? pydantic[email]
